# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/tasks": {
            "put": {
                "description": "Creates new message task that will be consumed.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task State"
                ],
                "summary": "Updates Task State",
                "parameters": [
                    {
                        "description": "Add TaskStateUpdateRequest",
                        "name": "TaskStateUpdateRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TaskStateUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.TaskStateUpdateResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/tasks/messages": {
            "get": {
                "description": "Gets message tasks with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message Task"
                ],
                "summary": "Gets message tasks with pagination",
                "parameters": [
                    {
                        "enum": [
                            "COMPLETED",
                            "WAITING",
                            "PROCESSING",
                            "FAILED"
                        ],
                        "type": "string",
                        "description": "status filter",
                        "name": "status",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "number",
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maximum": 20,
                        "minimum": 1,
                        "type": "number",
                        "description": "page size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PageResponse-model_MessageTask"
                        }
                    }
                }
            }
        },
        "/api/v1/tasks/messages/": {
            "post": {
                "description": "Creates new message task that will be consumed.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message Task"
                ],
                "summary": "CreatesMessageTask",
                "parameters": [
                    {
                        "description": "MessageTaskCreateRequest",
                        "name": "MessageTaskCreateRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.MessageTaskCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.MessageTaskCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/tasks/messages/{id}": {
            "get": {
                "description": "Gets message task result from cache",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message Task"
                ],
                "summary": "Gets message task result from cache",
                "parameters": [
                    {
                        "type": "string",
                        "description": "message id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.MessageTaskResult"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes message task that will",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message Task"
                ],
                "summary": "Deletes Message Task",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "message task id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "response ok",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "health",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.MessageTaskCreateRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "dto.MessageTaskCreateResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.PageResponse-model_MessageTask": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.MessageTask"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                }
            }
        },
        "dto.TaskStateUpdateRequest": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                }
            }
        },
        "dto.TaskStateUpdateResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "model.MessageTask": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/model.TaskStatus"
                },
                "to": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.MessageTaskResult": {
            "type": "object",
            "properties": {
                "messageId": {
                    "type": "string"
                },
                "sendingTime": {
                    "type": "string"
                },
                "taskId": {
                    "type": "string"
                }
            }
        },
        "model.TaskStatus": {
            "type": "string",
            "enum": [
                "WAITING",
                "PROCESSING",
                "COMPLETED",
                "FAILED"
            ],
            "x-enum-varnames": [
                "WAITING",
                "PROCESSING",
                "COMPLETED",
                "FAILED"
            ]
        }
    }
}